requirements.txt
fastapi
uvicorn
faiss-cpu
sentence-transformers
pydantic
pydantic-settings>=2.0
dotenv
pytest
httpx